test: split order_merge_test into part1/part2 to satisfy domain test file-size gate (#3404)
The order merge unify work introduced order_merge_test.dart at 431 physical
lines, exceeding the 400-line domain-package test file limit enforced by
repo.domain_package_test_file_size. Split into two focused files (moves/
diplomacy/build/work/research and naval/trade/multi-player/ordering); behavior
and coverage unchanged.
